content/html/content/test/test_bug143220.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=143220
-->
<head>
  <title>Test for Bug 143220</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=143220">Mozilla Bug 143220</a>
<p id="display">
  <input type="file" id="i1">
  <input type="file" id="i2">
</p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 143220 **/
var leafName;
var fullPath;

function initVals() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  var dirSvc = Components.classes["@mozilla.org/file/directory_service;1"]
                         .getService(Components.interfaces.nsIProperties);
  var file = dirSvc.get("XpcomLib", Components.interfaces.nsILocalFile);
  isnot(file, null, "Must have file here");

  leafName = file.leafName;
  fullPath = file.path;
}

function initControl1() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalFileRead");
  $("i1").value = fullPath;
  is($("i1").value, fullPath, "Should have set full path 1");
}

function initControl2() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  $("i2").value = fullPath;
  is($("i2").value, fullPath, "Should have set full path 2");
}

initVals();

// Check that we can't just set the value
try {
  $("i1").value = fullPath;
  is(0, 1, "Should have thrown exception on set!");
} catch(e) {
  is($("i1").value, "", "Shouldn't have value here");
}

initControl1();
initControl2();

is($("i1").value, leafName, "Leaking full value?");
is($("i2").value, leafName, "Leaking full value?");

</script>
</pre>
</body>
</html>
